Added testing for Discover Page
Created test functions using @Test from JetPack Compose. The code verifies that the page loads correctly, that the "create new community" button works, that you can join/leave any community and that you can see all the communities. Waiting for idle and thread.sleep() are esential in these functions, because the ui seems to need a bit more time to update correctly, due to the fact that we need to acces community_users to get the communities in which the users is a member.